Move: 移动脚本路径和命名空间
