+++ Cyan-Stars/Assets/Scripts/Gameplay/MusicGame/UI/ScoreSettlement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e3fef988f957b048b0d7a1179c2492b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
