+++ Cyan-Stars/Assets/Scripts/Gameplay/MusicGame/UI/ScoreSettlement/TmpAutoWidthBox.cs
@@ -12,7 +12,7 @@ using UnityEngine;
 using TMPro;
 using UnityEngine.UI;
 
-namespace CyanStars.Utils
+namespace CyanStars.Gameplay.MusicGame.UI.ScoreSettlement
 {
     [ExecuteAlways]
     [DisallowMultipleComponent]
@@ -87,6 +87,8 @@ namespace CyanStars.Utils
 
         private void AdjustWidth()
         {
+            Debug.Log("1");
+
             // 1. 先给一个极大的宽度，让 AutoSize 能够完全根据高度来计算出正确的字体大小
             rectTransform.SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, 999999f);
